Add result to pageant's command argument
When using '-c', the command is executed wether an SSH key has been
successfully added or not. If for example the user clicks 'Cancel'
on the password prompt, the command will still be executed, probably
resulting in a lot of user-facing errors due to the missing SSH key.

This commit adds an argument to the command that is either "0" or
"the number of SSH keys that have not been successfully added".
That way the called program can behave accordingly.

This is a very special customization for a specific usecase.